Another small additions to the list
* Clarified 0-chan's engine name, since it works on its own fork of Meguca
instead of upstream one.

* Added 0chanru.net, another 0chan-like imageboard with a long history
and short posting, which recently was resurrected.

* Added links to 8kun's Lokinet mirrors, as well as additional mirrors
of Endchan.

* Added Erlach, a WebSocket-based SPA imageboard, links to Rfch mirrors,
and an additional mirror of 0chan.pl.

* Added 102chan, an I2P-only imageboard.

* Added Metator. It was already present as an overchan in additional
chan link list, hovewer, this time its /meta/ board is being added. As
it was said before, the site lacks maintenance, as well as having no
formal rules, so illegal and offensive content might be posted from time
to time without being deleted. This is rarely the case, though, but if
you are not comfortable with that, feel free to remove it from the list
afterwards.

* Added Torodvach, a 2ch.ru spoof in Tor and I2P networks. "Torodvach"
isn't its official name, as it is referred simply as "Dvach" on all
pages, however, due to high ambiguity of this name, "Torodvach" got
widely adopted. The same lackluster of rules and control from Metator
might be a case here too, so again, be wary.